Changed PieChart to Doughnut Chart
Changed PieChart to Doughnut Chart
