+++ src/views/BarChart.vue
@@ -1,7 +1,13 @@
 <template>
   <div class="chart">
-    <h3>Tasks For Each Month</h3>
-    <chart></chart>
+    <h4>Tasks For Each Month</h4>
+    <!-- <div id="no-data"> 
+      <br>
+      <h5>You do not have any task...</h5>
+    </div> -->
+    <!-- <div id="data"> -->
+      <chart></chart>
+    <!-- </div> -->
   </div>
 </template>
 
@@ -10,7 +16,7 @@ import Chart from "./BarChart.js";
 
 export default {
   components: {
-    Chart
+    Chart,
   }
 };
 </script>
@@ -20,5 +26,7 @@ export default {
   text-align: center;
   align-content: center;
   margin: auto;
+  /* height: 440px;
+  width:440px; */
 }
 </style>
