+++ src/views/Dashboard.vue
@@ -7,7 +7,7 @@
       <LineChart></LineChart>
     </div>
     <div id="chart">
-      <PieChart></PieChart>
+      <Doughnut></Doughnut>
     </div>
   </div>
 </template>
@@ -26,7 +26,7 @@
   /* height:240px;
   width:240px;  */
   border-style: solid;
-  border-color:rgb(155, 84, 84);
+  border-color:#C9CBCF;
   border-width: 1px;
   padding:30px;
   float:left;
