+++ src/views/LineChart.js
@@ -9,21 +9,17 @@ export default {
         taskList: "",
         groups:"",
         datacollection: {
-            labels: [],
+            labels: ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sept", "Oct", "Nov", "Dec"],
             datasets: [
-              { 
-                data: [],
-                label: "",
-                borderColor: "#3e95cd",
-                fill: false
-              }
+              // { 
+              //   data: [],
+              //   label: "",
+              //   borderColor: "#3e95cd",
+              //   fill: false
+              // }
           ]
         },
         options: {
-            title: {
-              display: true,
-              //text: 'My Contribution'
-            },
             scales: {
               yAxes: [{
                 ticks: {
@@ -38,26 +34,13 @@ export default {
   },
 
   methods: {
-    getMonth: function(timestamp) {
-      let monthNumber = timestamp.toDate().getMonth();
-      let months = {
-        "0": "Jan",
-        "1": "Feb",
-        "2": "Mar",
-        "3": "Apr",
-        "4": "May",
-        "5": "Jun",
-        "6": "Jul",
-        "7": "Aug",
-        "8": "Sep",
-        "9": "Oct",
-        "10": "Nov",
-        "11": "Dec"
-      };
-      let month = months[monthNumber];
-      return month;
-    },
 
+    getColor: function() {
+      let colors = ["#C9CBCF", "#9966FE", "#36A2EB", "#4BC0C0", "#FFC134", "#FF6383"];
+      let randomNumber = Math.round(Math.random() * 5);
+      return colors[randomNumber];
+    },
+  
     fetchTasks: function() {
       let currUser = this.user;
       let tasks = {};
@@ -68,11 +51,9 @@ export default {
         myModules.forEach(function(module) {
           for (let groups in module.data()) {
             if (groups == "NoGroup") {
-              //groups = group name
               return;
             }
             if (!module.data()[groups]["Group Members"].includes(currUser)) {
-              //filter to show only user's groups
               return;
             }
             myGroups["members"][module.id] = module.data()[groups]["Group Members"];
@@ -82,19 +63,16 @@ export default {
         });
         this.taskList = tasks;
         this.groups = myGroups;
-        // console.log(this.taskList)
-        // console.log(this.groups)
 
         let selectedModule = "";
-        let numOfModules = Object.keys(tasks).length;
         let month = "";
         let num = "";
-        let count = 0;
         for (let module in tasks) {
           selectedModule = module;
-          for (let task in tasks[module]){
-            month = this.getMonth(tasks[module][task]["deadline"]);
-            if (tasks[module][task]["completed"] == true) {
+          let taskArray = [];
+          for (let task in tasks[selectedModule]){
+            month = tasks[selectedModule][task]["deadline"].toDate().getMonth();
+            if (tasks[selectedModule][task]["completed"] == true) {
               if (month in time) {
                 num = time[month];
                 time[month] = num + 1;
@@ -104,11 +82,9 @@ export default {
             } 
           }
           for (month in time){
-            this.datacollection.labels.push(month);
-            this.datacollection.datasets[count].data.push(time[month]);
-            this.datacollection.datasets[count].label = selectedModule;
+            taskArray[month] = time[month];
           }
-          count++;
+          this.datacollection.datasets.push({data:taskArray, label: selectedModule, borderColor: this.getColor(),fill:false})
         }
         this.renderChart(this.datacollection, this.options)
       });
